requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
imagehash==4.1.0          # via -r requirements.in
jinja2==2.11.2            # via -r requirements.in
markupsafe==1.1.1         # via jinja2
numpy==1.19.1             # via imagehash, pywavelets, scipy
pillow==7.2.0             # via -r requirements.in, imagehash, reportlab
pywavelets==1.1.1         # via imagehash
reportlab==3.5.46         # via -r requirements.in
scipy==1.5.2              # via imagehash
six==1.15.0               # via imagehash
tqdm==4.48.2              # via -r requirements.in
